pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fast-zero"
version = "0.0.0"
description = "Código do curso FastAPI do Zero"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
alembic = "^1.13.2"
fastapi = "^0.111"
pwdlib = {version = "^0.2.0", extras = ["argon2"]}
pydantic = {version = "^2.8.2", extras = ["email"]}
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.31"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.3.1"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
httpx = "^0.27.0"

[tool.ruff]
target-version = "py312"
line-length = 119
src = ["fast_zero", "tests"]

[tool.ruff.format]
line-ending = "lf"
indent-style = "space"
quote-style = "single"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.pytest.ini_options]
addopts = "--strict-config --cov-report=term-missing --no-cov-on-fail --cov=fast_zero"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
